requirements.txt
requests
black==19.10b0
eth-brownie == 1.18.1
flake8==3.7.9
isort==4.3.21
pre-commit==2.4.0
tox==3.15.1
